Moved the complete rbl handling in its own file. Rewritten the rbl code so that it is possible to have more configurtion possibilities. The rbllist file is now looking like this: basedomain       action              matchon          Rejectmessage relays.ordb.org  [reject|addheader]  [any|127.0.0.2]  See ordb.org
The fields are seperated by tabs (not spaces!!).
action defaults to reject but matchon defaults not to any!
The rest (environment RBL and RBLONLYHEADER, etc.) have not changed.

While doing that move the multiple times used replace function into the str.a
library. The function is now called byte_repl and the prototype is in byte.h.